requirements.txt
flask==3.0.3
google-generativeai==0.8.2
python-dotenv==1.0.1
markdown==3.7
